WiP on refactoring;

+++ tests/Aer.ConsistentHash.Tests/HashRingTests.cs
@@ -166,20 +166,23 @@ public class HashRingTests
     }
 
     [DataTestMethod]
-    [DataRow(5, (uint)1, 2)]
-    [DataRow(5, (uint)2, 3)]
-    [DataRow(5, (uint)3, 4)]
-    [DataRow(5, (uint)4, 5)]
-    [DataRow(5, (uint)5, 5)]
-    [DataRow(5, (uint)10, 5)]
-    [DataRow(5, (uint)0, 1)]
+    [DataRow(5, 1U, 2)]
+    [DataRow(5, 2U, 3)]
+    [DataRow(5, 3U, 4)]
+    [DataRow(5, 4U, 5)]
+    [DataRow(5, 5U, 5)]
+    [DataRow(5, 10U, 5)]
+    [DataRow(5, 0U, 1)]
     public void GetNodes_WithReplication(int nodesNumber, uint replicationFactor, int totalNodes)
     {
         var hashRing = GetHashRing();
 
         var nodesToAdd = Enumerable.Range(0, nodesNumber).Select(i => new Node()).ToArray();
+
         hashRing.AddNodes(nodesToAdd);
-        var nodes = hashRing.GetNodes(new[] { "test" }, replicationFactor: replicationFactor);
+
+        var nodes = 
+            hashRing.GetNodes(new[] {"test"}, replicationFactor: replicationFactor);
 
         nodes.Count.Should().Be(totalNodes);
         foreach (var node in nodes)
